refactor: table cell base class
