+++ lib/contentful_converter/nodes/tables/table.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-
 module ContentfulConverter
   module Nodes
     module Tables
