+++ lib/contentful_converter/nodes/tables/table_data.rb
@@ -1,36 +1,16 @@
 # frozen_string_literal: true
 
-require 'contentful_converter/nodes/base'
-
+require_relative './table_cell_base'
 
 module ContentfulConverter
   module Nodes
     module Tables
-      class TableData < Base
+      class TableData < TableCellBase
         def needs_p_wrapping?
           false
         end
 
-        def to_h(params = {})
-          h = super
-          h[:content] = [empty_node] if params[:content].empty?
-          h
-        end
-
         private
-
-        def empty_node
-          empty_node =  {
-            :nodeType => "paragraph",
-            :data => {},
-            :content => [{
-              :nodeType => "text",
-              :data => {},
-              :value => "",
-              :marks => []
-            }]
-          }
-        end
 
         def type
           'table-cell'
